core/fetch: Fix a bug where not all items in a fetch were persisted.
The bug was that the high watermark was being updated as the items in
the latest fetch were being processed. Thus, if the feed is ordered with
the newest item first, we would stop persisting the items after that,
even if they were newer than the database's high watermark.

The solution is simply to compare against the database high watermark
and only update this with the maximum seen watermark among all items
in the fetch.