core/fetch: Support pausing and restarting the fetcher.
This is needed to support adding or removing a feed via the admin
interface. It is too much of a hassle to try to do this with an
already-running fetcher since we'd need to selectively stop and start
some sub goroutines. Instead, we just stop them all, do the operation
on the database, and restart it, which will fetch all feeds again.

The latter half of this will come in a later commit.