core/db: Remove unnecessary read-after-write on imports.
Now that CockroachDB 1.1+ supports `ON CONFLICT ... RETURNING`, we can
just get the generated ID in the same insert call.

The current approach does needless updates on conflicts in order to
actually appear to have affected a row to get back the ID, but this is
not a big deal as imports are done only rarely.