core, frontend: Support local execution for Mercury Parser API.
The hosted version is deprecated. This commit supports invoking the CLI
for the parser directly from the goliath binary to do parsing.

Since the CLI is just part of a Javascript library, adding it as a dep
to the frontend is the simplest way to make it part of the toolchain.