core/fetcher: Don't unescape content if escaped quotes are found.
This was the root cause for why alt-text was being truncated on XKCD
entries when they had double-quotes in them. The RSS feed returns
HTML-escaped "&quot;" values which get heuristically unescaped. This
then appears as an actual closing quote character for the entire
attribute, making all text thereafter not match any rule and get
stripped.

This change introduces the small possibility that an article with only
HTML-escaped single and double quote characters will not be unescaped.