core/cache: Redirect unauthenticated cache request to proxied URL.
Instead of returning HTTP Code 407, which is an error, we now instead
URL-decode the requested URL and return a redirect to it. This means
that the client transparently falls back to retrieving the URL directly
(typically via HTTP) instead of not showing anything.

This mechanism won't provide any additional privilege for
unauthenticated requests that the client does not already have since the
second request is made directly from the client.