core: Add support for multiple users.
This commit introduces a per-user database structure with an updated schema and corresponding code. The user ID is now a primary key prefix of all tables (which are now all interleaved at the top-level under the user table).

Also adds some basic latency logging for slow DB operations.

The admin server also now requires a username to be specified when adding a new feed.