core/cache: Add new retrieval cache that to keep track of retrieved articles.
This class uses a cuckoo filter as a probabilistic way of determining if an article has already been fetched. This is necessary for the edge case when a read article is already GC'd but still present on a remote feed that does not publish timestamps: we currently do not realize that the article has been fetched in the past and re-fetch it since the lack of a timestamp prevents us from knowing that it is an old entry.