core/fetch: Support de-duplicating very similar articles from past fetches.
This is to address the case when a feed makes multiple edits to an already-published article and re-publishes it. Previously, this would appear as a separate entry since the hashes would not match and the published date would be newer.

With the default flag value, if a newly fetched article has the same link and at least 90% of the title and summary are equal to an older and unread article, the older one is removed in favor of the newest version.