frontend: Cache image previews in ArticleList instead of generating in entries.
This avoids regenerations of image previews every time the entry's props change.

While the images themselves are cached in the browser, this still wastes effort in
parsing the article HTML and initiating an image fetch very often.

The new cache in ArticleList persists across prop changes and has a max size instead.
It is lazily populated as entries are rendered at any point in the ArticleList
lifecycle.