backend/db: Add retries for opening the connection to the DB.
This is needed because with a Dockerized deployment, CockroachDB is brought up just before the main Goliath binary is started, which can lead to race conditions when trying to initialize a connection before the DB is ready to serve.

There is no trivial way to wait for the DB to be healthy from just within the DB's container since common utilities like `curl(1)` are unavailable in the deployment image.