frontend: Refactor data types into a new `models` folder.
This change also introduces class versions of each of the models to encapsulate operations on them (but are not used yet).

The purpose of this is to make it easier for callers to make tree changes as part of the initial retrieval or mark operations. This will benefit both `App.tsx` and FetchAPI implementations.