frontend: Add plumbing for switching to class-based data structures.
This change implements most of what's needed to switch to the class-based data types except for using the new classes in the child components.

This change as-is is not good to push since it currently consumes twice the memory to keep effectively two copies of client state around. The next changes will start actually using the new class-based structures after which the original copies can be removed.