backend/middleware: Allow serving request from the static directory without cookie verification.
There are some assets (like the web manifest, favicons, etc.) that need to be served when the user is not logged in.

Since all the content is from an `http.FileServer` rooted at the directory of static files, this should not be affected by path traversal attacks.

Cookie verification is still enabled when doing HTTPS reverse image proxying since all of those requests should come from authenticated users.