backend/storage: Base64 encode the retrieval cache before persisting to the DB.
This addresses a bug where the retrieval cache could not persisted if it contained invalid UTF-8 code points.

This was previously hex-encoded but that was omitted during a recent refactor. Still, this is not great. The data type of the column should instead be BYTES which is better for this binary data.