backend/cache: Fix decoding issue when initializing retrieval caches.
At startup, we fetch the persisted retrieval caches. For legacy reasons, this is stored as a string type (this should be changed to BYTE at some point). During initialization, we were failing to decode the base64 encoding before trying to recreate the retrieval cache.

This prevented using retrieval caches across restarts and resulted in spurious article resurrections.

Additionally, return a fatal error when unable to initialize even empty caches for all users as this will cause all future retrieval caches operations to be invalid and cause frequent resurrections.